Generic Button component moved to common folder
