+++ src/components/loadFromFile/ImageSizeDisplay.jsx
